All tests in a final dir working
