+++ tests/integration-e2e/test_connection.py
@@ -1,18 +1,14 @@
-import os
 import pytest
 from pathlib import Path
 from neurodamus.io.synapse_reader import SynapseParameters
 from neurodamus.node import Node
 
 
-SIM_DIR = Path(__file__).parent.absolute() / "simulations" / "v5_sonata"
+SIM_DIR = Path(__file__).parent.parent.absolute() / "simulations" / "v5_sonata"
 CONFIG_FILE_MINI = SIM_DIR / "simulation_config_mini.json"
 
 
 @pytest.mark.slow
-@pytest.mark.skipif(
-    not os.environ.get("NEURODAMUS_NEOCORTEX_ROOT"),
-    reason="Test requires loading a neocortex model to run")
 # This test is to mimic the error reported in HPCTM-1687 during connection.add_syanpses()
 # when detecting conn._synapse_params with more than one element is not None
 def test_add_synapses():
