+++ tests/integration/conftest.py
@@ -1,0 +1,6 @@
+import pytest
+# !! NOTE: Please don't import neron/neurodamus at module level
+# pytest weird discovery system will trigger Neuron init and open can of worms
+
+# Make all tests run forked
+pytestmark = pytest.mark.forked
