+++ tests/integration/test_cell.py
@@ -7,9 +7,6 @@ from pathlib import Path
 # !! NOTE: Please don't import neron/neurodamus at module level
 # pytest weird discovery system will trigger Neuron init and open can of worms
 
-# Make all tests run forked
-pytestmark = pytest.mark.forked
-
 
 @pytest.fixture(scope="session")
 def morphologies_root(rootdir):
