+++ tox.ini
@@ -35,14 +35,13 @@ commands =
     pytest -x --forked tests/integration
 
 
-[testenv:scientific]
+[testenv:e2e]
 # Please module load neurodamus-neocortex py-neurodamus beforehand
 passenv = *
 deps =
     pytest
     pytest-cov
     pytest-forked
-    pytest-xdist
 setenv =
     PYTHONPATH={toxinidir}:{env:PYTHONPATH}
     HOC_LIBRARY_PATH={toxinidir}/core/hoc:{env:HOC_LIBRARY_PATH}
@@ -52,7 +51,7 @@ allowlist_externals =
 commands =
     /usr/bin/echo {env:PYTHONPATH}
     /usr/bin/echo {env:HOC_LIBRARY_PATH}
-    pytest -s -x -n2 --forked --cov=neurodamus
+    pytest -s -x --forked {posargs:tests/integration-e2e}
 
 
 [testenv:flake8]
